requirements.txt
Django==3.0.2
django-extensions==2.2.6
ipython==7.12.0
djangorestframework==3.11.0
pandas==1.0.0
